ref: Split symbolication module into smaller pieces
This splits the main `symbolication/mod.rs` into smaller modules,
in particular this moves all the apple crash report, and symbolication
specific parts out of the main module which now concerns itself with
request dispatching.
However, each of the modules still includes way too much timeout and metrics-related code/wrappers. We might want to clean those up later as well.
